after fixing the responsive on the dashboard of the admin

+++ front-end/src/features/admin/Partenaires/adminPartnerService.js
@@ -3,25 +3,25 @@ import { protectedApi } from "../Login/authService";
 // Créer un nouveau partenaire (multipart/form-data)
 export const createPartnerAPI = async (partnerData) => {
     // partnerData doit être une instance de FormData
-    const res = await protectedApi.post(`/api/partenariats`, partnerData);
+    const res = await protectedApi.post(`/api/partenariats/admin`, partnerData);
     return res.data;
 };
 
 // Récupérer tous les partenaires
 export const fetchAdminPartnersAPI = async () => {
-    const res = await protectedApi.get(`/api/partenariats`);
+    const res = await protectedApi.get(`/api/partenariats/admin/all`);
     return res.data;
 };
 
 // Supprimer un partenaire
 export const deletePartnerAPI = async (id) => {
-    const res = await protectedApi.delete(`/api/partenariats/${id}`);
+    const res = await protectedApi.delete(`/api/partenariats/admin/${id}`);
     return res.data;
 };
 
 
 export const fetchPartnerByIdAPI = async (id) => {
-    const res = await protectedApi.get(`/api/partenariats/${id}`);
+    const res = await protectedApi.get(`/api/partenariats/admin/${id}`);
     return res.data;
 };
